+++ zapisy/apps/enrollment/courses/models/points.py
@@ -1,5 +1,18 @@
+"""Module points deals with values of the courses.
+
+It would seem, that just having a field `credits` in CourseEntity model would do
+the trick. Unfortunately, some courses have different value for different
+students, depending on the program they are pursuing (BSc, MSc) and their
+previous achievements.
+"""
+
+from typing import List, Optional, Dict, Iterable
+
+from django.conf import settings
 from django.db import models
-from django.db.models import Sum
+
+from apps.enrollment.courses.models.course import Course, CourseEntity
+from apps.enrollment.courses.models.semester import Semester
 from apps.users.models import Student
 
 
@@ -66,45 +79,108 @@ class PointsOfCourseEntities(models.Model):
         return '%s: %s %s' % (self.entity.name, self.value, self.type_of_point)
 
 
-class StudentPointsView(models.Model):
-    value = models.SmallIntegerField()
-    student = models.OneToOneField(Student, primary_key=True, on_delete=models.CASCADE)
-    entity = models.ForeignKey('courses.CourseEntity', on_delete=models.CASCADE)
+class StudentPointsView:
+    """Provides functions for counting ECTS points for a particular student.
 
-    class Meta:
-        managed = False
-        app_label = 'courses'
-
-    @classmethod
-    def get_student_points_in_semester(cls, student, semester):
-        """
-
-        Return sum of points in certain semester
-
-        @param student: users.Student object
-        @param semester: coruses.Semester object
-        @return: Integer
-        """
-        from apps.enrollment.records.models import Record
-
-        records = Record.enrolled.filter(
-            student=student, group__course__semester=semester).values_list(
-            'group__course__entity_id', flat=True).distinct()
-
-        return cls.get_points_for_entities(student, records)
+    A course does not necessarily always carry the same number of ECTS points.
+    It may earn differently depending on student's programme and the courses he
+    has already passed. The functions here implement this logic.
+    """
 
     @classmethod
-    def get_points_for_entities(cls, student, entities):
+    def student_points_in_semester(cls, student: Student, semester: Semester,
+                                   additional_courses: List[Course] = []) -> int:
+        """Computes sum of points in a semester from student's perspective.
+
+        Apart from the courses, the student is already enrolled into, it counts
+        in additional courses, in which he may not be present.
+
+        This function may give wrong historic result for a student who has
+        passed a certain course (like 'dyskretna_l') in the meantime.
+        """
+        from apps.enrollment.records.models import Record, RecordStatus
+        records = Record.objects.filter(
+            student=student, group__course__semester=semester,
+            status=RecordStatus.ENROLLED).values_list(
+                'group__course__entity_id', flat=True).distinct()
+        all_courses = list(set(list(records) + [c.entity_id for c in additional_courses]))
+        return cls.points_for_entities_total(student, all_courses)
+
+    @classmethod
+    def course_value_for_student(cls, student: Optional[Student], entity_id: int) -> int:
+        """Computes the value (number of ECTS credits) of a given course for a
+        student.
+        """
+        return cls.points_for_entities_total(student, [entity_id])
+
+    @classmethod
+    def points_for_entities_total(cls, student: Optional[Student], entity_ids: List[int]) -> int:
+        """Computes sum of points of entities from a student's perspective.
+
+        This function may give wrong historic result for a student who has
+        passed a certain course (like 'dyskretna_l') in the meantime.
+        """
+        return sum(cls.points_for_entities(student, entity_ids).values())
+
+    @classmethod
+    def points_for_entities(cls, student: Optional[Student], entity_ids: List[int]) -> Dict[int, int]:
+        """Computes points of entities from a student's perspective.
+
+        This function may give wrong historic result for a student who has
+        passed a certain course (like 'dyskretna_l') in the meantime.
+
+        The computation will be performed by obtaining all
+        PointsOfCourseEntities records for these courses and choosing the right
+        one for each course separately. For every entity in the list we see, if
+        it is one of the few special cases (compulsory courses with two levels).
+        If not, the PointsOfCourseEntities object should match the users program
+        or be None. The implementation looks crude, but it only performs a
+        constant number of database queries.
+
+        If the student is None, the function will return the default number of
+        credits for the course.
+
+        The returned Dict will be keyed by CourseEntity identifier.
         """
 
-        Return sum of student points for records
+        def value_with_program(
+                program_id: Optional[int],
+                points_of_courseentities_list: Iterable[PointsOfCourseEntities]) -> int:
+            """For a given program_id will find either the number of points
+            associated with this program_id, or with None, if one does not
+            exist.
+            """
+            poc: PointsOfCourseEntities
+            if program_id is not None:
+                for poc in points_of_courseentities_list:
+                    if poc.program_id == program_id:
+                        return poc.value
+            # The program_id is not on the list.
+            for poc in points_of_courseentities_list:
+                if poc.program_id is None:
+                    return poc.value
+            return 0
 
-        @param student:
-        @param records: Entity Id's list
-        @return:
-        """
-        points = cls.objects.\
-            filter(student=student, entity__in=entities).\
-            aggregate(Sum('value'))
-        # If there's no active term, value__sum will be None
-        return points['value__sum'] or 0
+        if student is None:
+            student_program_id = None
+        else:
+            student_program_id = student.program_id
+        entities = CourseEntity.objects.filter(
+            pk__in=entity_ids).prefetch_related('pointsofcourseentities_set')
+        points_per_entity: Dict[int, int] = dict()
+        entity: CourseEntity
+        for entity in entities:
+            # If the student had passed one of the BSc-level obligatory courses,
+            # the corresponding MSc-level course is worth as much for him as it
+            # would be for an MSc student.
+            bsc_courses = ["numeryczna_l", "dyskretna_l", "algorytmy_l", "programowanie_l"]
+            program_id = student_program_id
+            for bsc_course in bsc_courses:
+                # If student is None, getattr(student, 'attr', None) will also
+                # be None.
+                if getattr(entity, bsc_course) and getattr(student, bsc_course, None):
+                    program_id = settings.M_PROGRAM
+                    break
+            points_per_entity[entity.id] = value_with_program(
+                program_id, entity.pointsofcourseentities_set.all())
+        return points_per_entity