+++ zapisy/zapisy/settings.py
@@ -200,6 +200,7 @@ INSTALLED_APPS = (
     'pipeline',
     'apps.enrollment.courses',
     'apps.enrollment.records',
+    'apps.enrollment.timetable',
     'apps.statistics',
     'apps.news',
     'apps.offer.preferences',
@@ -244,14 +245,18 @@ LOGIN_REDIRECT_URL = '/users/'
 
 TEST_RUNNER = 'django.test.runner.DiscoverRunner'
 
-# settings for enrollment
-# ECTS_BONUS * ECTS = abs(t0-t1); set to 7, if changed, change also get_t0_interval()
+# Settings for enrollment.
+# Bonus minutes per one ECTS credit. This setting affects T0 times computation.
 ECTS_BONUS = 5
-ECTS_LIMIT = 35
+# Limits concerning the amount of ECTS points a student can sign up to in a
+# semester. For the first part of enrollment cycle, the INITIAL_LIMIT holds.
+# Then, after abolition time, students can enroll into some additional courses.
+ECTS_INITIAL_LIMIT = 35
 ECTS_FINAL_LIMIT = 45
 
 VOTE_LIMIT = 60
 
+# MSc Computer Science Program will have id=1 in database table users_program.
 M_PROGRAM = 1
 LETURE_TYPE = '1'
 QUEUE_PRIORITY_LIMIT = 5